Move session data functions into new file
