+++ webapp/app/forms/flows/multistep_flow.py
@@ -1,15 +1,13 @@
 from collections import namedtuple
-import zlib
 from typing import Optional
 
-from cryptography.fernet import InvalidToken
-from flask import redirect, request, url_for, session, json, abort
+from flask import redirect, request, url_for, session, abort
 from wtforms import Form
 
 from app import app
 
 # The RenderInfo is provided to all templates
-from app.crypto.encryption import decrypt, encrypt
+from app.forms.session_data import deserialize_session_data, override_session_data
 from app.forms.steps.step import FormStep
 
 
@@ -48,31 +46,6 @@ FlowNavItem = namedtuple(
     field_names=['number', 'text', 'active'],
     defaults=[False]
 )
-
-
-def serialize_session_data(data):
-    json_bytes = json.dumps(data).encode()
-    compressed = zlib.compress(json_bytes)
-    encrypted = encrypt(compressed)
-
-    return encrypted
-
-
-def deserialize_session_data(serialized_session, ttl: Optional[int] = None):
-    session_data = {}
-    if serialized_session:
-        try:
-            decrypted = decrypt(serialized_session, ttl)
-            decompressed = zlib.decompress(decrypted)
-            session_data = json.loads(decompressed.decode())
-        except InvalidToken:
-            app.logger.warning("Session decryption failed", exc_info=True)
-            session_data = {}
-    return session_data
-
-
-def override_session_data(stored_data):
-    session['form_data'] = serialize_session_data(stored_data)
 
 
 class MultiStepFlow:
