+++ webapp/tests/utils.py
@@ -4,7 +4,7 @@ from werkzeug.datastructures import ImmutableMultiDict
 
 from app import app
 from app.data_access.user_controller import create_user, find_user
-from app.forms.flows.multistep_flow import serialize_session_data
+from app.forms.session_data import serialize_session_data
 from app.forms.flows.step_chooser import StepChooser
